Add files via upload
I am currently working on handwritten digit recognize system where model will train on one data set and test it on the other one. By doing this, the accuracy should be over 98%. 
I have done the reading the data set into 2d arrays but if someone can contribute to this will be appreciated. Come with an algorithm that can get accuracy above 98%